refactor: extract function to parse-tx.ts for uint test
